move viam button to viam_example_widgets
