+++ example/viam_direct_machine_connection_example_app/android/app/build.gradle
@@ -45,7 +45,7 @@ android {
         applicationId "com.example.viam_example_app"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
-        minSdkVersion 23
+        minSdkVersion flutter.minSdkVersion
         targetSdkVersion flutter.targetSdkVersion
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
