+++ example/viam_example_widgets/lib/src/resources/audio_in.dart
@@ -3,7 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import 'package:viam_sdk/viam_sdk.dart';
-import 'package:viam_sdk/widgets.dart';
+
+import '../button.dart';
 
 /// A widget to capture audio from an [AudioIn] device.
 ///
