+++ example/viam_example_widgets/lib/src/resources/encoder.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
 import 'package:viam_sdk/viam_sdk.dart';
-import 'package:viam_sdk/widgets.dart';
+
+import '../button.dart';
 
 /// A widget to display and reset the position of an [Encoder].
 class ViamEncoderWidget extends StatefulWidget {
