+++ example/viam_example_widgets/lib/src/resources/generic.dart
@@ -3,7 +3,8 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 
 import 'package:viam_sdk/viam_sdk.dart';
-import 'package:viam_sdk/widgets.dart';
+
+import '../button.dart';
 
 /// A widget to send `doCommand` requests to a [Generic] component.
 ///
