+++ example/viam_example_widgets/lib/src/resources/gripper.dart
@@ -5,6 +5,8 @@ import 'package:flutter/widgets.dart';
 import 'package:viam_sdk/viam_sdk.dart';
 import 'package:viam_sdk/widgets.dart';
 
+import '../button.dart';
+
 /// A widget to control a [Gripper].
 ///
 /// This widget provides open and grab functionality for a [Gripper],
