+++ example/viam_example_widgets/lib/src/resources/vision.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 
 import 'package:viam_sdk/protos/service/vision.dart' as vision_pb;
 import 'package:viam_sdk/viam_sdk.dart';
-import 'package:viam_sdk/widgets.dart';
+
+import '../button.dart';
 
 /// A widget to run a vision service against a camera.
 ///
