+++ example/viam_example_widgets/lib/viam_example_widgets.dart
@@ -1,3 +1,4 @@
+export 'src/button.dart';
 export 'src/joystick.dart';
 export 'src/refreshable_data_table.dart';
 export 'src/resources/arm.dart';
