+++ example/viam_example_widgets/test/refreshable_data_table_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:viam_example_widgets/viam_example_widgets.dart';
-import 'package:viam_sdk/widgets.dart';
 
 import 'test_utils.dart';
