+++ lib/widgets.dart
@@ -1,3 +1,2 @@
-export 'widgets/button.dart';
 export 'widgets/camera_stream.dart';
 export 'widgets/multi_camera_stream.dart';
